Confirmation alert fo deleting an insertion.

+++ src/main/java/it/unipi/dii/largescale/secondchance/connection/controller/InsertionListController.java
@@ -98,9 +98,11 @@ public class InsertionListController {
         );
 
         delete.setOnMouseClicked(event -> {
-            ConnectionMongoDB.connMongo.deleteInsertionMongo(id);
-            ConnectionNeo4jDB.connNeo.deleteInsertionNeo4J(id);
-            initialize(Session.getLogUser().getUsername());
+            if(Utility.confirmDeletion()) {
+                ConnectionMongoDB.connMongo.deleteInsertionMongo(id);
+                ConnectionNeo4jDB.connNeo.deleteInsertionNeo4J(id);
+                initialize(Session.getLogUser().getUsername());
+            }
         });
 
         GridPane.setHalignment(image, HPos.LEFT);
